+++ e2e/tests/editor.e2e.js
@@ -1,5 +1,7 @@
 import { INLINE_TOOLBAR_BUTTONS } from '../cypress/dataHooks';
 
+/* eslint-disable mocha/no-skipped-tests */
+
 describe('editor', () => {
   beforeEach(() => cy.switchToDesktop());
 
@@ -9,24 +11,42 @@ describe('editor', () => {
         'Leverage agile frameworks',
         'to provide a robust synopsis for high level overviews.',
       ])
+      .setSelection(0, 0)
       .blurEditor()
       .matchSnapshots();
   });
 
-  it('should allow to apply inline styles', () => {
+  it('should allow to apply inline styles and links', () => {
     cy.loadEditor('plain')
-      .setTextStyle(INLINE_TOOLBAR_BUTTONS.BOLD, [0, 5])
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.BOLD, [40, 10])
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.UNDERLINE, [10, 5])
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.ITALIC, [20, 5])
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.BOLD, [30, 5])
+      .setLineSpacing(1, [10, 50])
+      .setColor(4, [200, 208])
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.UNDERLINE)
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.ITALIC)
-      .setTextStyle(INLINE_TOOLBAR_BUTTONS.ALIGNMENT)
-      .setTextStyle(INLINE_TOOLBAR_BUTTONS.TEXT_ALIGN_CENTER)
+      .setAlignment(INLINE_TOOLBAR_BUTTONS.TEXT_ALIGN_CENTER)
+      .setAlignment(INLINE_TOOLBAR_BUTTONS.TEXT_ALIGN_RIGHT)
+      .setAlignment(INLINE_TOOLBAR_BUTTONS.TEXT_ALIGN_LEFT)
       .setTextStyle(INLINE_TOOLBAR_BUTTONS.QUOTE, [30, 170])
-      .setColor(3, [20, 30])
-      .setLineSpacing(3, [30, 100])
+      .setLink([0, 10], 'https://www.wix.com/')
+      .setLink([50, 65], 'https://www.one.co.il/')
+      .setColor(1, [300, 305])
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.TITLE, [250, 260])
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.QUOTE, [250, 260])
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.ORDERED_LIST)
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.UNORDERED_LIST)
+      .setLineSpacing(3, [100, 150])
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.CODE_BLOCK, [100, 300])
+      .setLink([15, 30], 'https://www.sport5.co.il/')
       .setSelection(0, 0)
+      .enterParagraphs(['#LIVING THE DREAM\n'])
+      .setLink([0, 17], 'https://www.sport5.co.il')
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.CODE_BLOCK, [0, 10])
+      .enterParagraphs(['@NO_MORE\n'])
+      .setLink([0, 10], 'https://www.wix.com/')
+      .setTextStyle(INLINE_TOOLBAR_BUTTONS.CODE_BLOCK, [0, 10])
       .blurEditor()
       .matchSnapshots();
   });
@@ -43,6 +63,7 @@ describe('editor', () => {
       .alignImage('left')
       .alignImage('center')
       .alignImage('right')
+      .setSelection(0, 0)
       .matchSnapshots();
   });
 
@@ -52,10 +73,10 @@ describe('editor', () => {
     context('desktop', () => {
       beforeEach(() => cy.switchToDesktop());
 
-      it('should render plugin toolbar in rtl', () => {
+      it('should render side toolbar in rtl', () => {
         cy.loadEditor()
           .focusEditor()
-          .openPluginToolbar();
+          .openSideToolbar();
       });
 
       it('should render text toolbar in rtl', () => {