+++ packages/plugin-gallery/web/src/gallery-viewer.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { isEqual } from 'lodash';
+import { isEqual, get } from 'lodash';
 import { validate, mergeStyles, Context } from 'wix-rich-content-common';
 import { convertItemData } from './helpers/convert-item-data';
-import { getDefault } from './constants';
+import { getDefault, isHorizontalLayout } from './constants';
 import resizeMediaUrl from './helpers/resize-media-url';
 import schema from '../statics/data-schema.json';
 import viewerStyles from '../statics/styles/viewer.scss';
 import 'pro-gallery/dist/statics/main.min.css';
+import ExpandIcon from './icons/expand.svg';
 
 const { ProGallery } = process.env.SANTA ? {} : require('pro-gallery');
 
@@ -16,7 +17,10 @@ class GalleryViewer extends React.Component {
     validate(props.componentData, schema);
     super(props);
 
-    this.state = this.stateFromProps(props);
+    this.state = {
+      size: {},
+      ...this.stateFromProps(props),
+    };
 
     this.sampleItems = [1, 2, 3].map(i => {
       return {
@@ -33,25 +37,44 @@ class GalleryViewer extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
+    let galleryKey = this.state && this.state.galleryKey;
     if (!isEqual(nextProps.componentData, this.props.componentData)) {
+      const { galleryLayout: currentGalleryLayout } = this.props.componentData.styles;
+      const { galleryLayout: nextGalleryLayout } = nextProps.componentData.styles;
+      if (currentGalleryLayout !== nextGalleryLayout) {
+        this.updateDimensions(nextProps.componentData.styles);
+      }
       validate(nextProps.componentData, schema);
+      galleryKey = get(nextProps, 'componentData.styles.galleryLayout', Math.random());
     }
-    this.setState(this.stateFromProps(nextProps), () => this.updateDimensions());
+    this.setState({ galleryKey, ...this.stateFromProps(nextProps) });
   }
 
   componentDidMount() {
+    if (this.context.helpers.onExpand) {
+      const styleParams = this.state.styleParams;
+      this.setState({
+        styleParams: { ...styleParams, allowHover: true },
+      });
+    }
     this.updateDimensions();
-    window.addEventListener('resize', this.updateDimensions);
+    window.addEventListener('resize', this.handleResize);
   }
 
   componentWillUnmount() {
-    window.removeEventListener('resize', this.updateDimensions);
+    window.removeEventListener('resize', this.handleResize);
   }
 
-  updateDimensions = () => {
+  handleResize = () => this.updateDimensions();
+
+  updateDimensions = (styleParams = this.props.componentData.styles) => {
     if (this.container && this.container.getBoundingClientRect) {
       const width = Math.floor(this.container.getBoundingClientRect().width);
-      this.setState({ size: { width } });
+      let height;
+      if (isHorizontalLayout(styleParams)) {
+        height = width ? Math.floor((width * 3) / 4) : 300;
+      }
+      this.setState({ size: { width, height } });
     }
   };
 
@@ -62,10 +85,10 @@ class GalleryViewer extends React.Component {
       Object.assign(defaults.styles, props.componentData.styles || {}),
       this.hasTitle(items)
     );
-    // TODO remove gallery key
-    const galleryKey = Math.random();
+    if (this.context && this.context.helpers.onExpand) {
+      styleParams.allowHover = true;
+    }
     return {
-      galleryKey,
       items,
       styleParams,
     };
@@ -82,23 +105,28 @@ class GalleryViewer extends React.Component {
     }
   }
 
-  // handle pro-gallery events
-  // https://github.com/wix-incubator/pro-gallery/blob/master/packages/gallery/src/utils/constants/events.js
   handleGalleryEvents = (name, data) => {
     switch (name) {
-      // container size change callback
       case 'GALLERY_CHANGE':
-        this.container && (this.container.style.height = `${data.layoutHeight}px`);
-        this.setState(prevState => ({
-          size: {
-            ...prevState.size,
-            height: data.layoutHeight,
-          },
-        }));
+        if (this.container) {
+          if (!isHorizontalLayout(this.state.styleParams)) {
+            this.container.style.height = `${data.layoutHeight}px`;
+          } else {
+            this.container.style.height = 'auto';
+          }
+        }
+        break;
+      case 'ITEM_ACTION_TRIGGERED':
+        this.handleExpand(data);
         break;
       default:
         break;
     }
+  };
+
+  handleExpand = data => {
+    const { onExpand } = this.context.helpers;
+    onExpand && onExpand(this.props.entityIndex, data.idx);
   };
 
   hasTitle = items => {
@@ -127,22 +155,40 @@ class GalleryViewer extends React.Component {
     };
   };
 
+  hoverElement = itemProps => {
+    return itemProps.linkData.url ? (
+      <ExpandIcon
+        className={this.viewerStyles.expandIcon}
+        onClick={e => {
+          e.preventDefault();
+          this.handleExpand(itemProps);
+        }}
+      />
+    ) : null;
+  };
+
   render() {
     this.styles = this.styles || mergeStyles({ styles: viewerStyles, theme: this.context.theme });
+    const { scrollingElement, ...settings } = this.props.settings;
     // TODO remove gallery key
     const { galleryKey, styleParams, size = { width: 300 } } = this.state;
     const items = this.getItems();
     return (
-      <div ref={elem => (this.container = elem)} className={this.styles.gallery_container}>
+      <div
+        key={galleryKey}
+        ref={elem => (this.container = elem)}
+        className={this.styles.gallery_container}
+        data-hook="galleryViewer"
+      >
         <ProGallery
-          // TODO remove gallery key
-          key={galleryKey}
           items={items}
           styles={styleParams}
           container={size}
-          settings={this.props.settings}
+          settings={settings}
+          scrollingElement={scrollingElement}
           eventsListener={this.handleGalleryEvents}
           resizeMediaUrl={resizeMediaUrl}
+          customHoverRenderer={this.hoverElement}
         />
       </div>
     );
@@ -151,6 +197,7 @@ class GalleryViewer extends React.Component {
 
 GalleryViewer.propTypes = {
   componentData: PropTypes.object.isRequired,
+  entityIndex: PropTypes.number.isRequired,
   onClick: PropTypes.func,
   className: PropTypes.string,
   settings: PropTypes.object,